+++ src/oakutils/tools/display.py
@@ -1,7 +1,8 @@
-from typing import Optional, Dict, Union, Iterable, Tuple
+from typing import Optional, Dict, Union, Iterable, Tuple, Callable
 from threading import Thread
 import atexit
 import time
+from collections import defaultdict
 
 import cv2
 import numpy as np
@@ -48,6 +49,7 @@ class DisplayManager:
 
     def __init__(self, fps: int = 15, display_size: Tuple[int, int] = (640, 480)):
         self._displays: Dict[str, _Display] = {}
+        self._transforms: Dict[str, Callable] = defaultdict(lambda: lambda x: x)
         self._display_size = display_size
         self._fps = fps
         atexit.register(self._stop)
@@ -71,8 +73,21 @@ class DisplayManager:
             self._displays[name] = _Display(name, self._fps)
             self._displays[name](frame)
 
+    def set_transform(self, name: str, transform: Callable):
+        """
+        Sets a transform for the given name.
+
+        Parameters
+        ----------
+        name : str
+            The name of the transform
+        transform : Callable
+            The transform to set
+        """
+        self._transforms[name] = transform
+
     def update(
-        self, data: Union[Tuple[str, np.ndarray], Iterable[Tuple[str, np.ndarray]]]
+        self, data: Union[Tuple[str, np.ndarray], Iterable[Tuple[str, np.ndarray]]], transform: Optional[Callable] = None
     ):
         """
         Updates the display with the given data.
@@ -81,10 +96,19 @@ class DisplayManager:
         ----------
         data : Union[Tuple[str, np.ndarray], Iterable[str, np.ndarray]]
             The data to update the display with. Can be a single tuple or an iterable of tuples.
+        transform : Optional[Callable], optional
+            A transform to call on each frame, by default None
+            The transform should take in an np.ndarray and return an np.ndarray
         """
-        try:
+        # whether or not we are in Tuple or Iterable case
+        if isinstance(data, tuple):
             name, frame = data
-            self._update(name, frame)
-        except TypeError:
+            if transform is not None:
+                self.set_transform(name, transform)
+            self._update(name, self._transforms[name](frame))
+        else:
+            name, _ = data[0]
+            if transform is not None:
+                self.set_transform(name, transform)
             for name, frame in data:
-                self._update(name, frame)
+                self._update(name, self._transforms[name](frame))
